WIP: Upgrades Jetty 9.4.43.v20210629 to 10.0.15
Correct configuration of websocket compression. Compression is provided out of the box by Jetty's `permessage-deflate` extension.

Previously Openfire was registering the `permessage-deflate` extension, I assume this was attempting to enable websocket compression.

Presently, websocket compression is enabled by default in Jetty. So the correct way to control websocket compression is to disable the `permessage-deflate` extension when compression is not wanted:
https://github.com/eclipse/jetty.project/issues/1341